test(chatbot): fold tier3_test.go contents into commands_test.go
"Tier 3" was an internal triage label that doesn't belong in the
filename — future readers won't know what it means. The tests are
command tests like any other, so they live alongside helpCmd /
kilometresCmd / stateCmd in commands_test.go.

Also drops a stale comment about deferred GetScore-chain paths
(the chain variants landed in this same change).
